Information functions moved to .ExchangeData
